Move APE_TEST_CHAIN_ID to tests constants file.
