--- tests/acceptance/constants.py
@@ -1,3 +1,0 @@
-from ape.utils import DEFAULT_TEST_CHAIN_ID
-
-APE_TEST_CHAIN_ID = DEFAULT_TEST_CHAIN_ID  # APE uses this chain id
